+++ demos/main/src/main/java/com/google/android/exoplayer2/demo/SampleChooserActivity.java
@@ -43,8 +43,8 @@ import android.widget.Toast;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import com.google.android.exoplayer2.MediaItem;
+import com.google.android.exoplayer2.MediaItem.ClippingConfiguration;
 import com.google.android.exoplayer2.MediaMetadata;
-import com.google.android.exoplayer2.ParserException;
 import com.google.android.exoplayer2.RenderersFactory;
 import com.google.android.exoplayer2.offline.DownloadService;
 import com.google.android.exoplayer2.upstream.DataSource;
@@ -53,6 +53,7 @@ import com.google.android.exoplayer2.upstream.DataSourceUtil;
 import com.google.android.exoplayer2.upstream.DataSpec;
 import com.google.android.exoplayer2.util.Log;
 import com.google.android.exoplayer2.util.Util;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
 import java.io.InputStream;
@@ -327,8 +328,7 @@ public class SampleChooserActivity extends AppCompatActivity
             reader.nextString(); // Ignore.
             break;
           default:
-            throw ParserException.createForMalformedManifest(
-                "Unsupported name: " + name, /* cause= */ null);
+            throw new IOException("Unsupported name: " + name, /* cause= */ null);
         }
       }
       reader.endObject();
@@ -351,6 +351,8 @@ public class SampleChooserActivity extends AppCompatActivity
       boolean drmSessionForClearContent = false;
       boolean drmMultiSession = false;
       boolean drmForceDefaultLicenseUri = false;
+      MediaItem.ClippingConfiguration.Builder clippingConfiguration =
+          new ClippingConfiguration.Builder();
 
       MediaItem.Builder mediaItem = new MediaItem.Builder();
       reader.beginObject();
@@ -367,10 +369,10 @@ public class SampleChooserActivity extends AppCompatActivity
             extension = reader.nextString();
             break;
           case "clip_start_position_ms":
-            mediaItem.setClipStartPositionMs(reader.nextLong());
+            clippingConfiguration.setStartPositionMs(reader.nextLong());
             break;
           case "clip_end_position_ms":
-            mediaItem.setClipEndPositionMs(reader.nextLong());
+            clippingConfiguration.setEndPositionMs(reader.nextLong());
             break;
           case "ad_tag_uri":
             mediaItem.setAdsConfiguration(
@@ -420,8 +422,7 @@ public class SampleChooserActivity extends AppCompatActivity
             reader.endArray();
             break;
           default:
-            throw ParserException.createForMalformedManifest(
-                "Unsupported attribute name: " + name, /* cause= */ null);
+            throw new IOException("Unsupported attribute name: " + name, /* cause= */ null);
         }
       }
       reader.endObject();
@@ -439,7 +440,8 @@ public class SampleChooserActivity extends AppCompatActivity
         mediaItem
             .setUri(uri)
             .setMediaMetadata(new MediaMetadata.Builder().setTitle(title).build())
-            .setMimeType(adaptiveMimeType);
+            .setMimeType(adaptiveMimeType)
+            .setClippingConfiguration(clippingConfiguration.build());
         if (drmUuid != null) {
           mediaItem.setDrmConfiguration(
               new MediaItem.DrmConfiguration.Builder(drmUuid)
@@ -463,13 +465,15 @@ public class SampleChooserActivity extends AppCompatActivity
               "drm_uuid is required if drm_force_default_license_uri is set.");
         }
         if (subtitleUri != null) {
-          MediaItem.Subtitle subtitle =
-              new MediaItem.Subtitle(
-                  subtitleUri,
-                  checkNotNull(
-                      subtitleMimeType, "subtitle_mime_type is required if subtitle_uri is set."),
-                  subtitleLanguage);
-          mediaItem.setSubtitles(Collections.singletonList(subtitle));
+          MediaItem.SubtitleConfiguration subtitleConfiguration =
+              new MediaItem.SubtitleConfiguration.Builder(subtitleUri)
+                  .setMimeType(
+                      checkNotNull(
+                          subtitleMimeType,
+                          "subtitle_mime_type is required if subtitle_uri is set."))
+                  .setLanguage(subtitleLanguage)
+                  .build();
+          mediaItem.setSubtitleConfigurations(ImmutableList.of(subtitleConfiguration));
         }
         return new PlaylistHolder(title, Collections.singletonList(mediaItem.build()));
       }