+++ library/core/src/main/java/com/google/android/exoplayer2/source/ProgressiveMediaSource.java
@@ -175,6 +175,8 @@ public final class ProgressiveMediaSource extends BaseMediaSource
       return this;
     }
 
+    @Deprecated
+    @Override
     public Factory setDrmSessionManager(@Nullable DrmSessionManager drmSessionManager) {
       if (drmSessionManager == null) {
         setDrmSessionManagerProvider(null);
@@ -184,6 +186,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
       return this;
     }
 
+    @Deprecated
     @Override
     public Factory setDrmHttpDataSourceFactory(
         @Nullable HttpDataSource.Factory drmHttpDataSourceFactory) {
@@ -194,6 +197,7 @@ public final class ProgressiveMediaSource extends BaseMediaSource
       return this;
     }
 
+    @Deprecated
     @Override
     public Factory setDrmUserAgent(@Nullable String userAgent) {
       if (!usingCustomDrmSessionManagerProvider) {