+++ RELEASENOTES.md
@@ -1,5 +1,34 @@
 # Release notes
 
+### 2.16.1 (2021-11-18)
+
+*   Core Library:
+    *   Fix track selection issue where overriding one track group did not
+        disable other track groups of the same type
+        ([#9675](https://github.com/google/ExoPlayer/issues/9675)).
+    *   Fix track selection issue where a mixture of non-empty and empty track
+        overrides is not applied correctly
+        ([#9649](https://github.com/google/ExoPlayer/issues/9649)).
+    *   Add protected method `DefaultRenderersFactory.getCodecAdapterFactory()`
+        so that subclasses of `DefaultRenderersFactory` that override
+        `buildVideoRenderers()` or `buildAudioRenderers()` can access the codec
+        adapter factory and pass it to `MediaCodecRenderer` instances they
+        create.
+    *   Propagate ICY header fields `name` and `genre` to
+        `MediaMetadata.station` and `MediaMetadata.genre` respectively so that
+        they reach the app via `Player.Listener.onMediaMetadataChanged()`
+        ([#9677](https://github.com/google/ExoPlayer/issues/9677)).
+    *   Remove null keys from `DefaultHttpDataSource#getResponseHeaders`.
+*   Extractors:
+    *   WAV: Add support for RF64 streams
+        ([#9543](https://github.com/google/ExoPlayer/issues/9543)).
+*   DASH:
+    *   Add parsed essential and supplemental properties to the `Representation`
+        ([#9579](https://github.com/google/ExoPlayer/issues/9579)).
+*   HLS:
+    *   Correctly populate `Format.label` for audio only HLS streams
+        ([#9608](https://github.com/google/ExoPlayer/issues/9608)).
+
 ### 2.16.0 (2021-11-04)
 
 *   Core Library: